+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 ## Current (in progress)
 
 - Add a job to bind Tabular API to its datasets [#3162](https://github.com/opendatateam/udata/pull/3162)
+- CSW ISO Harvest: add extra_configs.remote_url_prefix [#3157](https://github.com/opendatateam/udata/pull/3157)
+- DCAT harvest : Refactor DCAT extras handling in rdf parsing [#3054](https://github.com/opendatateam/udata/pull/3054). `accesRights`, `rights` and `license` are now in `extras["dcat"]` for both datasets and resources. `accessRights` can be infered for a dataset from its resources.
+
+## 9.2.3 (2024-10-14)
+
+- Add migration to delete duplicate resources due to ODS harvesting [#3158](https://github.com/opendatateam/udata/pull/3158)
+- Add discussion.posted_on in discussion sort choices [3168](https://github.com/opendatateam/udata/pull/3168)
 
 ## 9.2.2 (2024-10-08)